refactor: rename internal/ to shared/ (ui + lib split)
The internal/ directory shipped in the npm package (so it was never private)
and mixed a reusable UI atom with a pure helper under a misleading name:

- ToolbarButton.vue -> shared/ui/ToolbarButton.vue
- fitDiagram.ts     -> shared/lib/fitDiagram.ts

Pure git mv + import-path updates in DmnModeler.vue, plus the package.json
"files" field and CLAUDE.md references. No behaviour change.
