+++ components/DmnModeler.vue
@@ -115,8 +115,8 @@ import { onSlideEnter } from '@slidev/client'
 import { useDmn } from '../composables/useDmn'
 import { camundaEngine } from '../engines/camunda'
 import type { Engine } from '../engines/types'
-import { fitDiagram } from '../internal/fitDiagram'
-import ToolbarButton from '../internal/ToolbarButton.vue'
+import { fitDiagram } from '../shared/lib/fitDiagram'
+import ToolbarButton from '../shared/ui/ToolbarButton.vue'
 
 const margin = 5
 const containerWaitTimeout = 5000
