+++ package.json
@@ -47,7 +47,7 @@
     "components",
     "composables",
     "engines",
-    "internal",
+    "shared",
     "vite.config.ts"
   ],
   "sideEffects": [
