+++ shared/lib/fitDiagram.ts
